refactor: RedisConfig 위치 변경 (core-domain)
