--- api/api-booking/src/main/java/com/pgms/apibooking/config/RedisConfig.java
@@ -1,23 +1,0 @@
-package com.pgms.apibooking.config;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.data.redis.connection.RedisConnectionFactory;
-import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
-import org.springframework.data.redis.core.RedisTemplate;
-
-@Configuration
-public class RedisConfig {
-
-	@Bean
-	LettuceConnectionFactory redisConnectionFactory() {
-		return new LettuceConnectionFactory();
-	}
-
-	@Bean
-	RedisTemplate<String, String> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
-		RedisTemplate<String, String> template = new RedisTemplate<>();
-		template.setConnectionFactory(redisConnectionFactory);
-		return template;
-	}
-}
