+++ core/core-domain/src/main/java/com/pgms/coredomain/RedisConfig.java
@@ -1,15 +1,22 @@
-package com.pgms.coredomain.domain.member.redis;
+package com.pgms.coredomain;
 
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.data.redis.connection.RedisConnectionFactory;
 import org.springframework.data.redis.connection.lettuce.LettuceConnectionFactory;
+import org.springframework.data.redis.core.RedisTemplate;
 
 import lombok.extern.slf4j.Slf4j;
 
 @Slf4j
 @Configuration
 public class RedisConfig {
+	@Bean
+	RedisTemplate<String, String> redisTemplate(RedisConnectionFactory redisConnectionFactory) {
+		RedisTemplate<String, String> template = new RedisTemplate<>();
+		template.setConnectionFactory(redisConnectionFactory);
+		return template;
+	}
 
 	@Bean
 	public RedisConnectionFactory redisConnectionFactory() {
